Add some progressive (but temporary) values for holdspace.
Planetary Travel just allowed adding more and more holdspace and then
changing the ship type based on hold size.  We're eventually going to
have distinct ship types but allow the user to amass multiple ships.

So all of the current stuff is just temporary until we get that figured
out.