Refactor a new Magnate toplevel class
Make a new toplevel class for invoking the Magnate client.  This is
a level above the dispatcher class.  While dispatcher handles events
once the program is running, magnate handles setting up the system.

For now, the magnate class also handles all of the dataloading
functions.  In the future we may break this back out along with adding
in state persistence.  Howver that may be database backed so it may
become a lot more complex at the same time.