Rework the logging a bit
New logging conventions

There is a logging module, magnate.logging.  It contains an
application-wide log object, magnate.logging.log.  This log has the name
field set to the name of the application, stellarmagnate.magnate.

Each python module creates a mlog variable from the application log
which sets the mod field to the __name__ of the module.

Individual functions further create flog variables from the mlogs which
set the func field to the name of the function.  This means that any
logs from within a function should include information about both what
module and what function they were emitted in.