Flowchart mockup of the Login-UI
The Login UI is realy a sequence of forms that walk the user through
logging in, creating an account, and creating a new Player.  This
flowchart takes us from program startup to the point where the Main
Window can be drawn.  It includes mockups of the various forms
encountered along the way.